pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synctalk"
version = "1.0.0"
description = "Commercial-grade 2D lip-sync video generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "opencv-python>=4.8.0",
    "numpy>=1.23.5,<2.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "scipy>=1.10.0",
    "configargparse>=1.7",
    "onnxruntime>=1.16.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]
gpu = [
    "onnxruntime-gpu>=1.16.0",
]

[project.scripts]
synctalk = "synctalk.cli:main"

[tool.setuptools.packages.find]
include = ["synctalk*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
